🚚 Rename: 파일명 오타 수정(AcitivityListWithPagination -> ActivityListWithPagination)
