+++ src/app/(public)/(explore)/(main)/ui/SearchResultSection.tsx
@@ -1,5 +1,5 @@
 'use client';
-import ActivityListWithPagination from '@/app/(public)/(explore)/(main)/ui/AcitivityListWithPagination';
+import ActivityListWithPagination from '@/app/(public)/(explore)/(main)/ui/ActivityListWithPagination';
 import CategoryFilter from '@/features/activity/main/ui/category-filter/CategoryFilter';
 import { useSearchResult } from '@/features/activity/search/hooks/useSearchResult';
 import EmptyState from '@/shared/ui/empty-state/EmptyState';
